+++ receiver/hostmetricsreceiver/internal/scraper/filesystemscraper/internal/metadata/generated_metrics_v2.go
@@ -17,6 +17,7 @@ type MetricSettings struct {
 type MetricsSettings struct {
 	SystemFilesystemInodesUsage MetricSettings `mapstructure:"system.filesystem.inodes.usage"`
 	SystemFilesystemUsage       MetricSettings `mapstructure:"system.filesystem.usage"`
+	SystemFilesystemUtilization MetricSettings `mapstructure:"system.filesystem.utilization"`
 }
 
 func DefaultMetricsSettings() MetricsSettings {
@@ -26,6 +27,9 @@ func DefaultMetricsSettings() MetricsSettings {
 		},
 		SystemFilesystemUsage: MetricSettings{
 			Enabled: true,
+		},
+		SystemFilesystemUtilization: MetricSettings{
+			Enabled: false,
 		},
 	}
 }
@@ -144,12 +148,67 @@ func newMetricSystemFilesystemUsage(settings MetricSettings) metricSystemFilesys
 	return m
 }
 
+type metricSystemFilesystemUtilization struct {
+	data     pdata.Metric   // data buffer for generated metric.
+	settings MetricSettings // metric settings provided by user.
+	capacity int            // max observed number of data points added to the metric.
+}
+
+// init fills system.filesystem.utilization metric with initial data.
+func (m *metricSystemFilesystemUtilization) init() {
+	m.data.SetName("system.filesystem.utilization")
+	m.data.SetDescription("Fraction of filesystem bytes used.")
+	m.data.SetUnit("1")
+	m.data.SetDataType(pdata.MetricDataTypeGauge)
+	m.data.Gauge().DataPoints().EnsureCapacity(m.capacity)
+}
+
+func (m *metricSystemFilesystemUtilization) recordDataPoint(start pdata.Timestamp, ts pdata.Timestamp, val float64, deviceAttributeValue string, modeAttributeValue string, mountpointAttributeValue string, typeAttributeValue string) {
+	if !m.settings.Enabled {
+		return
+	}
+	dp := m.data.Gauge().DataPoints().AppendEmpty()
+	dp.SetStartTimestamp(start)
+	dp.SetTimestamp(ts)
+	dp.SetDoubleVal(val)
+	dp.Attributes().Insert(A.Device, pdata.NewAttributeValueString(deviceAttributeValue))
+	dp.Attributes().Insert(A.Mode, pdata.NewAttributeValueString(modeAttributeValue))
+	dp.Attributes().Insert(A.Mountpoint, pdata.NewAttributeValueString(mountpointAttributeValue))
+	dp.Attributes().Insert(A.Type, pdata.NewAttributeValueString(typeAttributeValue))
+}
+
+// updateCapacity saves max length of data point slices that will be used for the slice capacity.
+func (m *metricSystemFilesystemUtilization) updateCapacity() {
+	if m.data.Gauge().DataPoints().Len() > m.capacity {
+		m.capacity = m.data.Gauge().DataPoints().Len()
+	}
+}
+
+// emit appends recorded metric data to a metrics slice and prepares it for recording another set of data points.
+func (m *metricSystemFilesystemUtilization) emit(metrics pdata.MetricSlice) {
+	if m.settings.Enabled && m.data.Gauge().DataPoints().Len() > 0 {
+		m.updateCapacity()
+		m.data.MoveTo(metrics.AppendEmpty())
+		m.init()
+	}
+}
+
+func newMetricSystemFilesystemUtilization(settings MetricSettings) metricSystemFilesystemUtilization {
+	m := metricSystemFilesystemUtilization{settings: settings}
+	if settings.Enabled {
+		m.data = pdata.NewMetric()
+		m.init()
+	}
+	return m
+}
+
 // MetricsBuilder provides an interface for scrapers to report metrics while taking care of all the transformations
 // required to produce metric representation defined in metadata and user settings.
 type MetricsBuilder struct {
 	startTime                         pdata.Timestamp
 	metricSystemFilesystemInodesUsage metricSystemFilesystemInodesUsage
 	metricSystemFilesystemUsage       metricSystemFilesystemUsage
+	metricSystemFilesystemUtilization metricSystemFilesystemUtilization
 }
 
 // metricBuilderOption applies changes to default metrics builder.
@@ -167,6 +226,7 @@ func NewMetricsBuilder(settings MetricsSettings, options ...metricBuilderOption)
 		startTime:                         pdata.NewTimestampFromTime(time.Now()),
 		metricSystemFilesystemInodesUsage: newMetricSystemFilesystemInodesUsage(settings.SystemFilesystemInodesUsage),
 		metricSystemFilesystemUsage:       newMetricSystemFilesystemUsage(settings.SystemFilesystemUsage),
+		metricSystemFilesystemUtilization: newMetricSystemFilesystemUtilization(settings.SystemFilesystemUtilization),
 	}
 	for _, op := range options {
 		op(mb)
@@ -180,6 +240,7 @@ func NewMetricsBuilder(settings MetricsSettings, options ...metricBuilderOption)
 func (mb *MetricsBuilder) Emit(metrics pdata.MetricSlice) {
 	mb.metricSystemFilesystemInodesUsage.emit(metrics)
 	mb.metricSystemFilesystemUsage.emit(metrics)
+	mb.metricSystemFilesystemUtilization.emit(metrics)
 }
 
 // RecordSystemFilesystemInodesUsageDataPoint adds a data point to system.filesystem.inodes.usage metric.
@@ -190,6 +251,11 @@ func (mb *MetricsBuilder) RecordSystemFilesystemInodesUsageDataPoint(ts pdata.Ti
 // RecordSystemFilesystemUsageDataPoint adds a data point to system.filesystem.usage metric.
 func (mb *MetricsBuilder) RecordSystemFilesystemUsageDataPoint(ts pdata.Timestamp, val int64, deviceAttributeValue string, modeAttributeValue string, mountpointAttributeValue string, typeAttributeValue string, stateAttributeValue string) {
 	mb.metricSystemFilesystemUsage.recordDataPoint(mb.startTime, ts, val, deviceAttributeValue, modeAttributeValue, mountpointAttributeValue, typeAttributeValue, stateAttributeValue)
+}
+
+// RecordSystemFilesystemUtilizationDataPoint adds a data point to system.filesystem.utilization metric.
+func (mb *MetricsBuilder) RecordSystemFilesystemUtilizationDataPoint(ts pdata.Timestamp, val float64, deviceAttributeValue string, modeAttributeValue string, mountpointAttributeValue string, typeAttributeValue string) {
+	mb.metricSystemFilesystemUtilization.recordDataPoint(mb.startTime, ts, val, deviceAttributeValue, modeAttributeValue, mountpointAttributeValue, typeAttributeValue)
 }
 
 // Reset resets metrics builder to its initial state. It should be used when external metrics source is restarted,